refactor testControlSrv to be part of testing.go and add some additional comments on what the end to end test is accomplishing
